+++ backend/crypto_api/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CryptoApiConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'crypto_api'
